agent-pay-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.1.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
requests==2.31.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
httpx>=0.24.0,<0.25.0
email-validator==2.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
